requirements/production.txt
# Celery
celery[redis,librabbitmq]==4.2.0
kombu==4.2.0
eventlet==0.21.0
flower==0.9.2

# Utils
python-dateutil==2.6.1

# Django
django<1.12
django-extensions
django-allauth==0.32
django-autocomplete-light==3.2.9
django-bootstrap3==8.2.3
django-cleanup==0.4.2
django-debug-toolbar==1.8
django-filter==1.0.4
django-file-form==0.2.4
django-localflavor== 1.5.2
django-material==1.0.0
django-querysetsequence==0.7.2
django-tables2==1.10.0
django-sequences==2.2
django-storages==1.6
django-dbconn-retry==0.1.5
Pillow==5.2.0
django-pandas==0.5.1
django-import-export

# Django channels
channels==1.1.8
asgi-redis==1.4.3
asgiref==2.0.1
daphne==1.3.0


# Python
boto==2.49
luigi==2.6.2
paramiko==2.2.1
psycopg2==2.7.3
pymssql==2.1.3
raven==6.2.1
six==1.10.0
SQLAlchemy==1.2.0b2
uwsgi==2.0.15
django-guardian==1.4.9
django-money==0.13.2
openpyxl==2.5.3
django-redis==4.9.0
retrying==1.3.3
django-codemirror-widget==0.4.1
djangorestframework==3.8.2
Markdown==2.6.11
django-cors-headers==2.2.0
pycpfcnpj==1.5
django-codemirror-widget==0.4.1
numpy==1.15.1
pandas==0.23.4
sendgrid==5.6.0

# docs
django-rest-swagger==2.2.0

# Oauth
django-oauth-toolkit==1.1.2
rest-condition==1.0.3